Refactor to VideoContent.tsx, PlaylistContent.tsx, and VideoListComponentLevel.tsx to use customhooks. This reuses a lot of code between them.
ChannelActions.tsx and VideoActionsBar.tsx components created to reduce code reuse between VideoContent.tsx, PlaylistContent.tsx, and VideoListComponentLevel.tsx.

If a playlist is empty, comments and buttons no longer appear, the page just shows text explaining the situation.
