+++ src/pages/Home/Components/VideoListComponentLevel.tsx
@@ -1,36 +1,21 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import { Box, useTheme } from "@mui/material";
+import React, { useEffect, useRef, useState } from "react";
 import { useSelector } from "react-redux";
-import { RootState } from "../../../state/store.ts";
-import { Avatar, Box, Button, Typography, useTheme } from "@mui/material";
-import { useFetchVideos } from "../../../hooks/useFetchVideos.tsx";
-import LazyLoad from "../../../components/common/LazyLoad.tsx";
-import {
-  BottomParent,
-  NameContainer,
-  ProductManagerRow,
-  VideoCard,
-  VideoCardCol,
-  VideoCardContainer,
-  VideoCardName,
-  VideoCardTitle,
-  VideoContainer,
-  VideoUploadDate,
-} from "./VideoList-styles.tsx";
-import ResponsiveImage from "../../../components/ResponsiveImage.tsx";
-import { formatDate, formatTimestampSeconds } from "../../../utils/time.ts";
-import { Video } from "../../../state/features/videoSlice.ts";
-import { queue } from "../../../wrappers/GlobalWrapper.tsx";
-import { VideoCardImageContainer } from "./VideoCardImageContainer.tsx";
+import { useParams } from "react-router-dom";
 import { QTUBE_VIDEO_BASE } from "../../../constants/Identifiers.ts";
+import { useFetchVideos } from "../../../hooks/useFetchVideos.tsx";
+import { Video } from "../../../state/features/videoSlice.ts";
+import { RootState } from "../../../state/store.ts";
+import { queue } from "../../../wrappers/GlobalWrapper.tsx";
+import { VideoManagerRow } from "./VideoList-styles.tsx";
+import VideoList from "./VideoList.tsx";
 
 interface VideoListProps {
   mode?: string;
 }
+
 export const VideoListComponentLevel = ({ mode }: VideoListProps) => {
   const { name: paramName } = useParams();
-  const theme = useTheme();
-  const [isLoading, setIsLoading] = useState<boolean>(true);
 
   const firstFetch = useRef(false);
   const afterFetch = useRef(false);
@@ -38,18 +23,9 @@ export const VideoListComponentLevel = ({ mode }: VideoListProps) => {
     (state: RootState) => state.video.hashMapVideos
   );
 
-  const countNewVideos = useSelector(
-    (state: RootState) => state.video.countNewVideos
-  );
-  const userAvatarHash = useSelector(
-    (state: RootState) => state.global.userAvatarHash
-  );
-
   const [videos, setVideos] = React.useState<Video[]>([]);
 
-  const navigate = useNavigate();
-  const { getVideo, getNewVideos, checkNewVideos, checkAndUpdateVideo } =
-    useFetchVideos();
+  const { getVideo, checkAndUpdateVideo } = useFetchVideos();
 
   const getVideos = React.useCallback(async () => {
     try {
@@ -98,21 +74,15 @@ export const VideoListComponentLevel = ({ mode }: VideoListProps) => {
         }
       }
     } catch (error) {
-    } finally {
+      console.log(error);
     }
   }, [videos, hashMapVideos]);
-
-  const getVideosHandler = React.useCallback(async () => {
-    if (!firstFetch.current || !afterFetch.current) return;
-    await getVideos();
-  }, [getVideos]);
 
   const getVideosHandlerMount = React.useCallback(async () => {
     if (firstFetch.current) return;
     firstFetch.current = true;
     await getVideos();
     afterFetch.current = true;
-    setIsLoading(false);
   }, [getVideos]);
 
   useEffect(() => {
@@ -122,7 +92,7 @@ export const VideoListComponentLevel = ({ mode }: VideoListProps) => {
   }, [getVideosHandlerMount]);
 
   return (
-    <ProductManagerRow>
+    <VideoManagerRow>
       <Box
         sx={{
           width: "100%",
@@ -131,68 +101,8 @@ export const VideoListComponentLevel = ({ mode }: VideoListProps) => {
           alignItems: "center",
         }}
       >
-        <VideoCardContainer>
-          {videos.map((video: any, index: number) => {
-            const existingVideo = hashMapVideos[video.id + "-" + video.user];
-            let hasHash = false;
-            let videoObj = video;
-            if (existingVideo) {
-              videoObj = existingVideo;
-              hasHash = true;
-            }
-
-            let avatarUrl = "";
-            if (userAvatarHash[videoObj?.user]) {
-              avatarUrl = userAvatarHash[videoObj?.user];
-            }
-
-            if (
-              hasHash &&
-              (!videoObj?.videoImage || videoObj?.videoImage?.length < 50)
-            ) {
-              return null;
-            }
-
-            return (
-              <VideoCardCol key={videoObj.id}>
-                <VideoCard
-                  onClick={() => {
-                    navigate(`/video/${videoObj.user}/${videoObj.id}`);
-                  }}
-                >
-                  <VideoCardImageContainer
-                    width={266}
-                    height={150}
-                    videoImage={videoObj.videoImage}
-                    frameImages={videoObj?.extracts || []}
-                  />
-                  <VideoCardTitle>{videoObj.title}</VideoCardTitle>
-                  <BottomParent>
-                    <NameContainer>
-                      <Avatar
-                        sx={{ height: 24, width: 24 }}
-                        src={`/arbitrary/THUMBNAIL/${videoObj?.user}/qortal_avatar`}
-                        alt={`${videoObj.user}'s avatar`}
-                      />
-                      <VideoCardName>{videoObj.user}</VideoCardName>
-                    </NameContainer>
-
-                    {videoObj?.created && (
-                      <VideoUploadDate>
-                        {formatDate(videoObj.created)}
-                      </VideoUploadDate>
-                    )}
-                  </BottomParent>
-                </VideoCard>
-              </VideoCardCol>
-            );
-          })}
-        </VideoCardContainer>
-        <LazyLoad
-          onLoadMore={getVideosHandler}
-          isLoading={isLoading}
-        ></LazyLoad>
+        <VideoList videos={videos} />
       </Box>
-    </ProductManagerRow>
+    </VideoManagerRow>
   );
 };
